refactor: rename anything containing 'boilerplate' and remove debug element
