--- docker-compose.yml
@@ -1,16 +1,0 @@
-# version: "3.7"
-
-services:
-  nanosp:
-    image: ghcr.io/blooo-io/speculos:latest
-    volumes:
-      - ./bin:/speculos/apps
-      - ./src:/speculos/sources
-      - ./build:/speculos/build
-    ports:
-      - "5000:5000" # api
-      - "40000:40000" # apdu
-    environment:
-      - GDB_DIRECTORY_LIST="/speculos/sources/:/speculos/sources/common:/speculos/sources/common/ui"
-    command: "-d --model ${DEVICE:-nanox} /speculos/build/${BUILD_DEVICE:-nanox}/bin/app.elf --display headless --apdu-port 40000"
-    # Add `--vnc-password "<password>"` for macos users to use built-in vnc client.
