+++ tests/application_client/command_sender.py
@@ -157,7 +157,7 @@ class Errors(IntEnum):
 # pylint: disable=too-many-public-methods
 # pylint: disable=too-many-locals
 # pylint: disable=too-many-lines
-class BoilerplateCommandSender:
+class CommandSender:
     def __init__(self, backend: BackendInterface) -> None:
         self.backend = backend
