+++ tests/test_appname_cmd.py
@@ -1,6 +1,6 @@
 import pytest
-from application_client.boilerplate_command_sender import BoilerplateCommandSender
-from application_client.boilerplate_response_unpacker import (
+from application_client.command_sender import CommandSender
+from application_client.response_unpacker import (
     unpack_get_app_name_response,
 )
 
@@ -9,7 +9,7 @@ from application_client.boilerplate_response_unpacker import (
 @pytest.mark.active_test_scope
 def test_app_name(backend):
     # Use the app interface instead of raw interface
-    client = BoilerplateCommandSender(backend)
+    client = CommandSender(backend)
     # Send the GET_APP_NAME instruction to the app
     response = client.get_app_name()
     # Assert that we have received the correct appname
