+++ tests/test_appversion_cmd.py
@@ -1,6 +1,6 @@
 import pytest
-from application_client.boilerplate_command_sender import BoilerplateCommandSender
-from application_client.boilerplate_response_unpacker import (
+from application_client.command_sender import CommandSender
+from application_client.response_unpacker import (
     unpack_get_app_name_response,
 )
 from utils import verify_version
@@ -10,7 +10,7 @@ from utils import verify_version
 @pytest.mark.active_test_scope
 def test_app_version(backend):
     # Use the app interface instead of raw interface
-    client = BoilerplateCommandSender(backend)
+    client = CommandSender(backend)
     # Send the GET_APP_NAME instruction to the app
     response = client.get_app_version()
     # Assert that we have received the correct appname
