+++ tests/test_deploy_module.py
@@ -1,10 +1,10 @@
 import pytest
 
-from application_client.boilerplate_command_sender import (
-    BoilerplateCommandSender,
+from application_client.command_sender import (
+    CommandSender,
     Errors,
 )
-from application_client.boilerplate_response_unpacker import (
+from application_client.response_unpacker import (
     unpack_get_public_key_response,
 )
 from ragger.bip import calculate_public_key_and_chaincode, CurveChoice
@@ -17,7 +17,7 @@ from utils import navigate_until_text_and_compare, instructions_builder
 def test_deploy_module(
     backend, navigator, test_name, default_screenshot_path
 ):
-    client = BoilerplateCommandSender(backend)
+    client = CommandSender(backend)
     path = "m/1105/0/0/0/0/2/0/0"
     header_and_type = bytes.fromhex(
         "20a845815bd43a1999e90fbf971537a70392eb38f89e6bd32b3dd70e1a9551d7000000000000000a0000000000000064000000290000000063de5da700"
