+++ tests/test_verify_address.py
@@ -3,11 +3,11 @@ from bip32 import BIP32, HARDENED_INDEX
 from bip_utils import Bip32Slip10Ed25519
 from cryptography.hazmat.primitives.asymmetric.ed25519 import Ed25519PrivateKey
 from cryptography.hazmat.primitives import serialization
-from application_client.boilerplate_command_sender import (
-    BoilerplateCommandSender,
+from application_client.command_sender import (
+    CommandSender,
     Errors,
 )
-from application_client.boilerplate_response_unpacker import (
+from application_client.response_unpacker import (
     unpack_get_public_key_response,
 )
 from ragger.bip import calculate_public_key_and_chaincode, CurveChoice
@@ -23,7 +23,7 @@ def test_verify_address_confirm_legacy_path_accepted(
     test_name: str,
     default_screenshot_path: str,
 ):
-    client = BoilerplateCommandSender(backend)
+    client = CommandSender(backend)
     with client.verify_address(identity_index=0, credential_counter=0):
         scenario_navigator.address_review_approve()
 
@@ -36,7 +36,7 @@ def test_verify_address_confirm_legacy_path_accepted(
 def test_verify_address_confirm_new_path_accepted(
     backend, scenario_navigator, test_name, default_screenshot_path
 ):
-    client = BoilerplateCommandSender(backend)
+    client = CommandSender(backend)
     with client.verify_address(identity_index=0, credential_counter=0, idp_index=0):
         scenario_navigator.address_review_approve()
 
@@ -49,7 +49,7 @@ def test_verify_address_confirm_new_path_accepted(
 def test_verify_address_confirm_refused(
     backend, scenario_navigator, test_name, default_screenshot_path
 ):
-    client = BoilerplateCommandSender(backend)
+    client = CommandSender(backend)
     try:
         with client.verify_address(identity_index=0, credential_counter=0, idp_index=0):
             scenario_navigator.address_review_reject()
